requirements.txt
langchain
langchain-core

langchain-google-genai
google-generativeai
transformers

langchain-huggingface
huggingface-hub

python-dotenv
numpy

typedict
pydantic
pydantic[email]
